feat(orga): add disabled filter button on learner page
