+++ orga/app/components/organization-participant/learner-filters.hbs
@@ -1,0 +1,26 @@
+<PixFilterBanner
+  @title={{t "pages.organization-participants.filters.title"}}
+  class="participant-filter-banner hide-on-mobile"
+  aria-label={{t "pages.organization-participants.filters.aria-label"}}
+  @details={{t "pages.organization-participants.filters.participations-count" count=@learnersCount}}
+  @clearFiltersLabel={{t "pages.organization-participants.filters.actions.clear"}}
+  @onClearFilters={{@onResetFilter}}
+  @isClearFilterButtonDisabled={{this.isClearFiltersButtonDisabled}}
+>
+  <Ui::SearchInputFilter
+    @field="fullName"
+    @value={{@fullName}}
+    @placeholder={{t "pages.organization-participants.filters.type.fullName.placeholder"}}
+    @ariaLabel={{t "pages.organization-participants.filters.type.fullName.title"}}
+    @triggerFiltering={{@triggerFiltering}}
+  />
+  <Ui::MultiSelectFilter
+    @field="certificability"
+    @title={{t "pages.organization-participants.filters.type.certificability.label"}}
+    @onSelect={{@triggerFiltering}}
+    @selectedOption={{@certificabilityFilter}}
+    @options={{this.certificabilityOptions}}
+    @placeholder={{t "pages.organization-participants.filters.type.certificability.placeholder"}}
+    @emptyMessage=""
+  />
+</PixFilterBanner>
