+++ orga/app/components/organization-participant/learner-filters.js
@@ -1,0 +1,27 @@
+import Component from '@glimmer/component';
+import { inject as service } from '@ember/service';
+
+export default class LearnerFilters extends Component {
+  @service intl;
+
+  get certificabilityOptions() {
+    return [
+      {
+        value: 'not-available',
+        label: this.intl.t('pages.sco-organization-participants.table.column.is-certifiable.not-available'),
+      },
+      {
+        value: 'eligible',
+        label: this.intl.t('pages.sco-organization-participants.table.column.is-certifiable.eligible'),
+      },
+      {
+        value: 'non-eligible',
+        label: this.intl.t('pages.sco-organization-participants.table.column.is-certifiable.non-eligible'),
+      },
+    ];
+  }
+
+  get isClearFiltersButtonDisabled() {
+    return !this.args.fullName && this.args.certificabilityFilter.length === 0;
+  }
+}
