+++ orga/app/templates/authenticated/organization-participants/list.hbs
@@ -8,9 +8,8 @@
       @triggerFiltering={{this.triggerFiltering}}
       @onResetFilter={{this.resetFilters}}
       @fullName={{this.fullName}}
-      @onClickLearner={{this.goToLearnerPage}}
       @certificabilityFilter={{this.certificability}}
-      @certificabilityOptions={{this.certificabilityOptions}}
+      @onClickLearner={{this.goToLearnerPage}}
       @participationCountOrder={{this.participationCountOrder}}
       @sortByParticipationCount={{this.sortByParticipationCount}}
       @sortByLastname={{this.sortByLastname}}
